Now all tests from given suite are run

+++ run-jstests.py
@@ -37,6 +37,20 @@ from os.path import isfile, join
 from subprocess import run, TimeoutExpired, STDOUT, PIPE, DEVNULL
 from time import perf_counter
 from collections import OrderedDict
+import re
+
+
+def getTestsForSuite(suite, mongo_root):
+    cmd = test_binary[:]
+    cmd.append('--suites={}'.format(suite))
+    cmd.append('-n')
+    proc = run(cmd, stdout=PIPE, cwd=mongo_root)
+    out = proc.stdout.decode('utf-8').splitlines()
+
+    regex = re.compile(r"^jstests[\w/_]+.*js")
+    out = [join(mongo_root, line) for line in out if regex.search(line)]
+
+    return out
 
 if __name__ == '__main__':
     parser = ArgumentParser(
@@ -53,15 +67,14 @@ if __name__ == '__main__':
     args = parser.parse_args()
 
     test_dir = join(args.mongo_root, 'jstests', args.suite)
-    if args.tests:
-        tests = args.tests
-    else:
-        tests = [test for test in listdir(test_dir) if isfile(
-            join(test_dir, test)) and test.endswith('js')]
-
     test_binary = [join(args.mongo_root, 'buildscripts', 'resmoke.py')]
     test_args = ['--continueOnFailure', '--storageEngine=pmse',
                  '--suites={}'.format(args.suite), '--dbpath={}'.format(args.dbpath)]
+
+    if args.tests:
+        tests = args.tests
+    else:
+        tests = getTestsForSuite(args.suite, args.mongo_root);
 
     failed = []
     passed_warnings = OrderedDict()
@@ -73,6 +86,8 @@ if __name__ == '__main__':
         cmd = test_binary + test_args
         cmd.append(join(test_dir, test))
         print_output = False
+        skipped = False
+
         print('{} ...'.format(test).ljust(margin), end='', flush=True)
 
         start = perf_counter()
@@ -87,7 +102,11 @@ if __name__ == '__main__':
         else:
             out = proc.stdout.decode('utf-8')
             if proc.returncode == 0:
-                print('PASSED', end='')
+                if "No tests ran" in out:
+                    print('SKIPPED', end='')
+                    skipped = True
+                else:
+                    print('PASSED', end='')
             elif 'were skipped, 0 failed, 0 errored' in out:
                 print('PASSED WITH WARNINGS. Test exited with code {}'.format(
                     proc.returncode), end='')
@@ -102,7 +121,8 @@ if __name__ == '__main__':
             print('\t{0:.3f} [ms]'.format(elapsed_ms))
             if print_output:
                 print(out)
-            run('rm -r {}/job0'.format(args.dbpath), shell=True)
+            if not skipped:
+                run('rm -r {}/job0'.format(args.dbpath), shell=True)
 
     if not failed and not timeout:
         print('All tests passed')
